feat: move simple VLA to examples and remove simple world model
