--- neuracore/ml/algorithms/simple_vla/__init__.py
@@ -1,1 +1,0 @@
-"""Init."""
