--- neuracore/ml/algorithms/simple_world_model/__init__.py
@@ -1,1 +1,0 @@
-"""Init."""
