--- neuracore/ml/config/algorithm/simple_vla.yaml
@@ -1,11 +1,0 @@
-# @package _global_
-algorithm:
-  _target_: neuracore.ml.algorithms.simple_vla.simple_vla.SimpleVLA
-  hidden_dim: 512
-  cnn_output_dim: 64
-  language_output_dim: 64
-  fusion_output_dim: 128
-  num_layers: 3
-  lr: 1e-4
-  lr_backbone: 1e-5
-  weight_decay: 1e-4
