--- neuracore/ml/config/algorithm/simple_world_model.yaml
@@ -1,9 +1,0 @@
-# @package _global_
-algorithm:
-  _target_: neuracore.ml.algorithms.simple_world_model.simple_world_model.SimpleWorldModel
-  hidden_dim: 512
-  cnn_output_dim: 128
-  num_layers: 3
-  lr: 1e-4
-  lr_backbone: 1e-5
-  weight_decay: 1e-4
